RORDEV-1025 clean stale sessions in case of index based sessions (#191)
* RORDEV-1025 Clean stale sessions in case of index based sessions

* RORDEV-1025 Refactor session cleanup info in Kibana doc

The previous description for the 'sessions_cleanup_interval' key in 'kibana.yml' was extended to provide a table for the interval definitions. The table explicitly lists the shorthand and corresponding measure - seconds(s), minutes(m), hours(h), and days(d) - to ensure accurate and easy customization of the automatic cleanup procedures. This improves usability and comprehension for configuring the Kibana session cleanup intervals.